Working force armor cost mods

+++ MMOCoreORB/src/server/zone/objects/creature/commands/ForceArmor2Command.h
@@ -41,7 +41,7 @@ public:
 
 		// TODO: Force Rank modifiers.
 		int enhSkills = playerObject->numSpecificSkills(creo, "force_discipline_enhancements_");
-		int enhMod = enhSkills * .025;
+		float enhMod = enhSkills * .025;
 		int forceCost = param * (0.5 - enhMod);
 		if (playerObject->getForcePower() <= forceCost) { // Remove buff if not enough force.
 			Buff* buff = creo->getBuff(BuffCRC::JEDI_FORCE_ARMOR_2);
